requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django_oauth_toolkit==1.3.2
autopep8==1.5.2
